can't import hardware from ot3api, breaks importablility
